move hash functions to utils.py instead of separate hash_utils.py
