--- python/sglang/srt/mem_cache/hash_utils.py
@@ -1,40 +1,0 @@
-"""Lightweight hash utilities for KV cache key management.
-
-These functions are intentionally in a separate module to avoid pulling in
-heavy CUDA dependencies (hicache_storage → memory_pool_host → sgl_kernel)
-when only hash computation is needed.
-"""
-
-import hashlib
-from typing import List, Optional
-
-
-def get_hash_str(token_ids: List[int], prior_hash: Optional[str] = None) -> str:
-    hasher = hashlib.sha256()
-
-    if prior_hash:
-        hasher.update(bytes.fromhex(prior_hash))
-
-    for t in token_ids:
-        if isinstance(t, tuple):
-            # EAGLE bigram mode: hash both elements to uniquely identify the bigram
-            for elem in t:
-                hasher.update(elem.to_bytes(4, byteorder="little", signed=False))
-        else:
-            # Regular mode: single integer token
-            hasher.update(t.to_bytes(4, byteorder="little", signed=False))
-
-    return hasher.hexdigest()
-
-
-def hash_str_to_int64(hash_str: str) -> int:
-    """Convert SHA256 hex string to signed 64-bit integer for events.
-
-    Takes first 16 hex characters (64 bits) and converts to signed int64 range.
-    """
-    # Take first 16 hex chars to get 64-bit value
-    uint64_val = int(hash_str[:16], 16)
-    # Convert to signed int64 range [-2^63, 2^63-1]
-    if uint64_val >= 2**63:
-        return uint64_val - 2**64
-    return uint64_val
