+++ python/sglang/srt/mem_cache/radix_cache.py
@@ -62,7 +62,7 @@ from sglang.srt.mem_cache.evict_policy import (
     PriorityStrategy,
     SLRUStrategy,
 )
-from sglang.srt.mem_cache.hash_utils import get_hash_str, hash_str_to_int64
+from sglang.srt.mem_cache.utils import get_hash_str, hash_str_to_int64
 
 if TYPE_CHECKING:
     from sglang.srt.managers.schedule_batch import Req
